backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic-settings==2.5.2
httpx==0.27.2
playwright==1.47.0
openai==1.58.1
apscheduler==3.10.4
python-dotenv==1.0.1
tweepy==4.16.0
apify-client==2.5.0
